ingestion-worker/requirements.txt
# Core dependencies
requests
schedule
python-dotenv
newsapi-python
pinecone-client

# Testing dependencies
pytest
pytest-mock
pytest-cov
